Correct Lesson 3 - add components
