+++ src/App.css
@@ -40,42 +40,3 @@ main {
   display: flex;
   gap: 30px;
 }
-
-/** LIST **/
-
-.msgList {
-  padding: 12px 10px;
-  border-bottom: dashed 1px #000;
-  width: 30vw;
-  display: flex;
-  gap: 20px;
-  background-color: #cccccc44;
-}
-
-.msgList:nth-child(2n) {
-  background-color: #ccccccff;
-}
-
-/**** FORM ****/
-
-.addMsgForm {
-  display: flex;
-  flex-direction: column;
-  width: 30vw;
-  padding: 10px;
-  gap: 20px;
-}
-
-.addMsgForm label {
-  display: flex;
-  justify-content: space-between;
-  gap: 20px;
-}
-
-.addMsgForm input, .addMsgForm textarea {
-  width: 100%;
-}
-
-.addMsgForm input[type="submit"] {
-  width: 50%;
-}
