+++ src/components/chatList/chatList.js
@@ -1,0 +1,28 @@
+// import React from 'react';
+
+// import List from '@material-ui/core/List';
+// import ListItem from '@material-ui/core/ListItem';
+// import ListItemText from '@material-ui/core/ListItemText';
+
+// export default function ChatList(props) {
+//     // const { list, curId, setChat } = props;
+
+//     // const handleChangeChat = (id) => {
+//     //     setChat(id);
+//     // }
+
+//     return (
+//         // <List subheader="Мои чаты:">
+//         //     { list.map((chat) => {
+//         //         return (
+//         //         <ListItem button 
+//         //                   key={ chat.id } 
+//         //                   selected={ chat.id === curId } 
+//         //                   onClick={ handleChangeChat(chat.id) }>
+//         //             <ListItemText primary={ chat.name }></ListItemText>
+//         //         </ListItem>
+//         //         );
+//         //     })}
+//         // </List>
+//     );
+// };
